Extract keyboard handlers into per-mode hooks
Splits the 200-line monolithic keydown handler in build.tsx into three
focused hooks: useSharedKeyboard (zoom, rotate, tab, shortcuts),
useSetupKeyboard (crop/polygon/arrows/undo), useBuildingKeyboard
(completion, annotations, step nav, timer). build.tsx drops from 368
to 167 lines.
